Moved room logic to room generator class
